+++ tests/test_extension_verification.py
@@ -1,5 +1,6 @@
 import re
 import pytest
+from gersemi.builtin_commands import builtin_commands
 from gersemi.extensions import verify, VerificationFailure
 from gersemi.extension_type import ModuleExtension
 
@@ -23,6 +24,8 @@ ab_signatures = f"{ab}['signatures']"
         {"ab": {"unsupported_entry": {}, "options": [], "foobar": {}}},
         {"ab": {"signatures": {"OK": {}}}},
         {"ab": {"signatures": {"OK": {"multi_value_keywords": "VALUES"}}}},
+        {"ab": {"options": ("ONE", "22", "THREE")}},
+        {"ab": {"one_value_keywords": ("ONE", "234", "THREE")}},
     ],
 )
 def test_extension_passes_verification(definition):
@@ -55,14 +58,6 @@ def test_extension_passes_verification(definition):
             f"{ab}['options']: keyword (2) has to be a string",
         ),
         (
-            {"ab": {"options": ("ONE", "22", "THREE")}},
-            f"{ab}['options']: keyword ('22') has to start with a letter or underscore",
-        ),
-        (
-            {"ab": {"one_value_keywords": ("ONE", "234", "THREE")}},
-            f"{ab}['one_value_keywords']: keyword ('234') has to start with a letter or underscore",
-        ),
-        (
             {"ab": {"front_positional_arguments": 34}},
             f"{ab}['front_positional_arguments']: is not a collection of strings (34)",
         ),
@@ -84,7 +79,7 @@ def test_extension_passes_verification(definition):
         ),
         (
             {"ab": {"sections": {"12": 34}}},
-            f"{ab_sections}: keyword ('12') has to start with a letter or underscore",
+            f"{ab_sections}['12']: is not a mapping",
         ),
         (
             {"ab": {"sections": {"ABC": 34}}},
@@ -101,7 +96,7 @@ def test_extension_passes_verification(definition):
         ),
         (
             {"ab": {"signatures": {"12": 34}}},
-            f"{ab_signatures}: signature ('12') has to start with a letter or underscore",
+            f"{ab_signatures}['12']: is not a mapping",
         ),
         (
             {"ab": {"signatures": {"CD": 34}}},
@@ -112,3 +107,15 @@ def test_extension_passes_verification(definition):
 def test_extension_fails_verification(definition, outcome):
     with pytest.raises(VerificationFailure, match=re.escape(outcome)):
         verify(ModuleExtension("foo"), definition)
+
+
+@pytest.mark.parametrize(
+    ["command", "definition"],
+    list(builtin_commands.items()),
+    ids=list(builtin_commands),
+)
+def test_extension_based_on_builtins_works(command, definition):
+    try:
+        verify(ModuleExtension("foo"), {command: definition})
+    except VerificationFailure as e:
+        pytest.fail(f"{repr(definition)} should pass verification but doesn't: {e}")